더 엄격한 요청 데이터에 대한 유효성 검사 수행

+++ src/api/components/api_schema.ts
@@ -68,7 +68,7 @@ export namespace APISchema {
     }
 
     export namespace Search {
-        export const cursor = z.coerce.number().optional();
+        export const cursor = z.coerce.number().int().optional();
         export const sort = z.enum(["oldest", "newest"]).default("newest");
     }
 }
